// informe-Comercial-final/Inf-Com-Client/src/app/models/formularioExpress.ts
export class FormExpress {
  _id: string;
  razonSocialExpress: string;
  fecha: string;
  solicitante: string;
  nombreComercialExpress: string;
  rucExpress: number;
  formaLegalExpress: string;
  estadoExpress: string;
  inicioActividadesExpress: string;
  actividadPrincipalExpress: string;
  provinciaCiudadExpress: string;
  direccionExpress: string;
  sucursalesExpress: string;
  telefonoExpress: number;
  paginaCorreoExpress: string;
  representanteLegalExpress: string;
  personaContactoExpress: string;
  cargoExpress: string;
  empresasRelacionadasExpress: string;
  numeroEmpleadosExpress: number;
  pagoRentaUnoExpress: string;
  pagoRentaDosExpress: string;
  impuestoDivisasExpressUno: string;
  impuestoDivisasExpressDos: string;
  //calificacion
  requeridoClienteExpress: string;
  requeridoServiratingExpress: string;
  nombreReferenciaExpress: string;
  direccionReferenciaExpress: string;
  telefonoReferenciaExpress: string;
  contactoReferenciaExpress: string;
  lineaCreditoExpress: string;
  pagosReferenciaExpress: string;
  productoServicioReferenciaExpress: string;
  clienteDesdeExpress: string;
  compraMensualExpress: string;
  //selects
  checks: String;
  checks1: String;
  puntualidad:String;
  //checks
  procesosJudicialesExpress: string;
  accionesDelitosExpress: string;
  informacionRelevanteExpress: string;
  informacionBancariaExpress: string;
  ventas2014Express: string;
  ventas2015Express:string;
  activos2014Express:string;
  activos2015Express: string;
  pasivo2014Express: string;
  pasivo2015Express: string;
  patrimonio2014Express: string;
  patrimonio2015Express: string;
  utilidad2014Express: string;
  utilidad2015Express:string;
  terrenoAreaExpress: string;
  terrenoAvaluoExpress: string;
  construccionAreaExpress: string;
  construccionAvaluoExpress: string;
  avaluoTotalAreaExpress: string;
  avaluoTotalAvaluoExpress: string;
  marcaExpress: string;
  marcaExpress1: string;
  marcaExpress2: string;
  marcaExpress3: string;
  marcaExpress4: string;
  tipoExpress: string;
  tipoExpress1: string;
  tipoExpress2: string;
  tipoExpress3: string;
  tipoExpress4: string;
  fechavahiculosExpress: string;
  fechavahiculosExpress1: string;
  fechavahiculosExpress2: string;
  fechavahiculosExpress3: string;
  fechavahiculosExpress4: string;
  avaluoExpress: string;
  avaluoExpress1: string;
  avaluoExpress2: string;
  avaluoExpress3: string;
  avaluoExpress4: string;
  observacionesExpress: string;


}